# docs/swagger.yaml
basePath: /api/v1
definitions:
  ginx.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  ginx.ResponseFail:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
  schema.LoginParam:
    properties:
      code:
        type: string
      password:
        type: string
      username:
        type: string
      uuid:
        type: string
    required:
    - code
    - password
    - username
    - uuid
    type: object
  schema.LoginTokenInfo:
    properties:
      access_expire:
        description: 过期时间戳
        type: integer
      access_token:
        description: 访问令牌
        type: string
      refresh_after:
        description: 刷新时间戳
        type: integer
    type: object
...
  schema.Notice:
    properties:
      content:
        type: string
      createdAt:
        type: string
      endTime:
        type: string
      id:
        type: integer
      operator:
        type: string
      startTime:
        type: string
      title:
        type: string
    type: object
  schema.NoticeCreateReq:
    properties:
      content:
        minLength: 2
        type: string
      endTime:
        type: string
      operator:
        maxLength: 150
        minLength: 2
        type: string
      startTime:
        type: string
      title:
        maxLength: 150
        minLength: 2
        type: string
    required:
    - content
    - endTime
    - operator
    - startTime
    - title
    type: object
  schema.NoticeGetPageResp:
    properties:
      list:
        items:
          $ref: '#/definitions/schema.Notice'
        type: array
      total:
        type: integer
    type: object
  schema.NoticeUpdateReq:
    properties:
      content:
        minLength: 2
        type: string
      endTime:
        type: string
      id:
        type: integer
      operator:
        maxLength: 150
        minLength: 2
        type: string
      startTime:
        type: string
      title:
        maxLength: 150
        minLength: 2
        type: string
    required:
    - content
    - endTime
    - id
    - operator
    - startTime
    - title
    type: object
...
  schema.SysApi:
    properties:
      action:
        description: 请求类型
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      deletedAt:
        description: 删除时间
        type: string
      handle:
        description: handle
        type: string
      id:
        description: 主键编码
        type: integer
      path:
        description: 地址
        type: string
      title:
        description: 标题
        type: string
      type:
        description: 接口类型
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
  schema.SysApiDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  schema.SysApiGetPageResp:
    properties:
      data:
        items:
          $ref: '#/definitions/schema.SysApi'
        type: array
      total:
        type: integer
    type: object
  schema.SysApiUpdateReq:
    properties:
      action:
        type: string
      handle:
        type: string
      id:
        type: integer
      path:
        type: string
      title:
        type: string
      type:
        type: string
      updateBy:
        type: integer
    type: object
...
host: 127.0.0.1:8000
info:
  contact:
    email: conjurer888888@gmail.com
    name: conjurer
    url: https:/github.com/dot123
  description: This is a game management background. you can use the api key `ApiKeyAuth`
    to test the authorization filters.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com
  title: gin-gorm-admin API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.LoginParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.LoginTokenInfo'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      summary: 登陆
      tags:
      - 认证
  /logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 登出
      tags:
      - 认证
...
  /msg/notice:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: pageNum
        type: integer
      - in: query
        maximum: 100
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.NoticeGetPageResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取公告列表
      tags:
      - 公告管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.NoticeCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 新建公告
      tags:
      - 公告管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.NoticeUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 更新公告
      tags:
      - 公告管理
  /msg/notice/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 公告id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 删除公告
      tags:
      - 公告管理
...
  /public/uploadFile:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      summary: 上传文件
      tags:
      - 文件管理
  /refresh_token:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.LoginTokenInfo'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 刷新token
      tags:
      - 认证
...
  /sys-api:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysApiDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 删除接口管理
      tags:
      - 接口管理
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: action
        type: string
      - in: query
        name: createdAtOrder
        type: string
      - in: query
        name: pageNum
        type: integer
      - in: query
        maximum: 100
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: string
      - in: query
        name: path
        type: string
      - in: query
        name: pathOrder
        type: string
      - in: query
        name: title
        type: string
      - in: query
        name: titleOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysApiGetPageResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取接口管理列表
      tags:
      - 接口管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysApiUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改接口管理
      tags:
      - 接口管理
  /sys-api/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysApi'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取接口管理
      tags:
      - 接口管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"